requirements.txt
pywebview, requests